pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "linescore"
version = "0.2.0"
description = "Code quality scoring via AI classification checks"
requires-python = ">=3.11"

[project.optional-dependencies]
anthropic = ["anthropic>=0.39.0"]
llamacpp = ["llama-cpp-python>=0.3.0", "huggingface-hub>=0.20.0"]
all = ["anthropic>=0.39.0", "llama-cpp-python>=0.3.0", "huggingface-hub>=0.20.0"]

[project.scripts]
linescore = "linescore.cli:main"
